chore: modify project structure to prepare for Azure Functions
