+++ coc-gsheet/config.ts.sample
@@ -1,0 +1,4 @@
+export default {
+  uri: "https://api.clashofclans.com/v1",
+  token: "enter your token here"
+};
