+++ coc-gsheet/db/clan.table.ts
