+++ coc-gsheet/index.ts
@@ -1,30 +1,25 @@
-import express, { Express, Request, Response } from "express";
 import { CocAPI } from "./api/coc_api";
 import { DbAPI } from "./db/db_api";
 import { PlayerController } from "./model/player/player.controller";
 import { ClanController } from "./model/clan/clan.controller";
-import { SpreadsheetAPI } from "../src/gsheet/spreadsheet_api";
+//import { Sequelize, DataTypes } from 'sequelize';
+import { SpreadsheetAPI } from "./gsheet/spreadsheet_api";
 import { PlayerSpreadsheet } from "./gsheet/player.data";
 
-const app: Express = express();
-
-app.get("/", async (req: Request, res: Response) => {
-  const docId: string = req.query.docId as string;
-  const clanTag: string = req.query.clanTag as string;
-  // TODO: validate input
-
+async function main() {
   // Init
   const api = new CocAPI(process.env["COC_TOKEN"]);
   const db = new DbAPI();
 
-  const clan = new ClanController("#" + clanTag, api, db);
+  // TODO: replace hardcoded clan tag
+  const clan = new ClanController("#20LCQ2CR8", api, db);
   await clan.get();
 
   if (clan.oClan.apiRetrieved) {
     clan.save();
 
     const gapi = new SpreadsheetAPI();
-    await gapi.loadDoc(docId);
+    await gapi.loadDoc("10cRyo1IOVh5fQmpuH34JnVGYIIRWhvPe27WLjyXpBrw"); // = TestAPI doc
     const playerSheet = new PlayerSpreadsheet(gapi.doc);
     await playerSheet.getSheet("API-PlayerData");
 
@@ -44,12 +39,6 @@ app.get("/", async (req: Request, res: Response) => {
       }
     }
   }
+}
 
-  res.send(
-    "Received: <UL><LI>docId: " + docId + "<LI>clanTag: " + clanTag + "</UL>"
-  );
-});
-
-app.listen(8000, () => {
-  console.log("Example app listening on port 8000!");
-});
+main();
