+++ coc-gsheet/model/clan/clan.controller.ts
