+++ coc-gsheet/model/clan/clan.model.ts
