+++ coc-gsheet/test/clanTestAPI.ts
@@ -1,7 +1,7 @@
 const assert = require("chai").assert;
-import { CocAPI } from "../src/api/coc_api";
-import { DbAPI } from "../src/db/db_api";
-import { ClanController } from "../src/model/clan/clan.controller";
+import { CocAPI } from "../api/coc_api";
+import { DbAPI } from "../db/db_api";
+import { ClanController } from "../model/clan/clan.controller";
 
 // TODO: move to specific prepare function?
 const api = new CocAPI(process.env.COC_TOKEN);
