+++ coc-gsheet/test/clanTestCOCAPI.ts
