+++ coc-gsheet/test/playerTestAPI.ts
@@ -34,9 +34,9 @@ decreases   decreasesBy     decreasesButNotBy
 //
 
 const assert = require("chai").assert;
-import { CocAPI } from "../src/api/coc_api";
-import { DbAPI } from "../src/db/db_api";
-import { PlayerController } from "../src/model/player/player.controller";
+import { CocAPI } from "../api/coc_api";
+import { DbAPI } from "../db/db_api";
+import { PlayerController } from "../model/player/player.controller";
 
 // TODO: move to specific prepare function?
 const api = new CocAPI(process.env.COC_TOKEN);
