+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "tsc",
     "dev": "npx tsc --watch",
-    "start": "node dist/src/server.js",
+    "start": "node dist/server.js",
     "test": "npx mocha 'dist/test/**/*.js' --recursive"
   },
   "keywords": [
@@ -18,9 +18,7 @@
   "author": "Stijn Mertens",
   "license": "ISC",
   "dependencies": {
-    "@types/express": "^4.17.17",
     "clashofclans.js": "^2.9.2",
-    "express": "^4.18.2",
     "google-spreadsheet": "^3.3.0",
     "node-fetch": "^3.1.0",
     "reflect-metadata": "^0.1.13",
@@ -28,12 +26,14 @@
   },
   "devDependencies": {
     "@types/chai": "^4.3.4",
+    "@types/express": "^4.17.17",
     "@types/google-spreadsheet": "^3.3.0",
     "@types/mocha": "^10.0.1",
     "@types/node": "^16.18.11",
     "@types/validator": "^13.7.10",
     "@typescript-eslint/eslint-plugin": "^5.44.0",
     "@typescript-eslint/parser": "^5.44.0",
+    "@azure/functions": "^3.5.0",
     "axios": "^1.2.0",
     "chai": "^4.3.7",
     "eslint": "^8.28.0",
